+++ MidtermEX1/Main.java
@@ -1,4 +1,4 @@
-package Midterm;
+package MidtermEX1;
     public class Main {
         public static void main(String[] args) {
             DiscreteSignal signal = new DiscreteSignal();
